moved VbProjectExtensions from reflection to extension namespace
