+++ RetailCoder.VBE/UnitTesting/NewUnitTestModuleCommand.cs
@@ -2,6 +2,7 @@
 using System.Runtime.InteropServices;
 using Microsoft.Vbe.Interop;
 using Rubberduck.Reflection;
+using Rubberduck.Extensions;
 
 namespace Rubberduck.UnitTesting
 {
